Complete script with testing
(refs #67)
