+++ python/terrain/__init__.py
@@ -1,0 +1,1 @@
+from .terrain import main
