+++ python/terrain/__main__.py
@@ -1,0 +1,4 @@
+import sys
+from .terrain import main
+if __name__ == '__main__':
+    sys.exit(main())
