+++ scripts/mesh_terrain.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+import os, sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '..', 'python'))
+import terrain
+sys.exit(terrain.main())
