--- scripts/python/terrain/__init__.py
@@ -1,1 +1,0 @@
-from .terrain import main
